requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
Pillow>=10.0.0
pypdfium2>=4.0.0
groq>=0.9.0
streamlit>=1.30.0
pandas>=2.0.0
